fix: game: moved Draw{Always,AsOneTimeFlag,OnlyWhen}() to global TU
Also removed unneeded randInt variable.
